Writing flow: extract arrow nav
